refactor: friend model #157
Friend model 을 friends app 으로 옮겼습니다
